libcamera: software_isp: debayer: Introduce a stop() callback to the debayer object
The eGL class wants to be able to teardown its sync_ data member
properly but, doing so in the destructor means we can't make the eGL
context current and thus can't tear down the sync primitive properly.

Introduce a stop() method to the debayer class which triggers from the
softisp's stop method, allowing a controlled and appropriate tear-down
of debayer-egl and egl class related data well before the destructors
get invoked.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>