libcamera: software_isp: debayer_egl: Implement input buffer import
Try to import the input buffer directly to avoid having to upload/copy
it. This is often limited by the buffer stride as GPUs tend to have high
alignment requirements, often 64, 128 or 256 bytes.

Also some cleanups to the DmaSyncer handling - we should only need to
care about the input buffer to be coherent on the CPU.

Signed-off-by: Robert Mader <robert.mader@collabora.com>