pipeline: simple: Prefer input configs with stride aligned to 64 bytes
In order to improve chances that a GPU dmabuf import succeeds. While
using 256 would essentially guarantee a matching stride with all vendors
- notably AMD - 64 seems to be enough for most mobile/embedded/low-end
devices while filtering out way fewer configs, making it a seemingly
good compromise.

Note that the actual stride used by v4l2 will usually be aligned - we
assume to 16 bytes for now as that again appears to be a good balance with
both low false-negatives and false-positives.

Signed-off-by: Robert Mader <robert.mader@collabora.com>